Finish backtest function and small refactors

+++ src/model/price_predictor.py
@@ -35,7 +35,7 @@ class PricePredictor:
 
         Returns
         -------
-        dict<int, float>
+        dict<int, int>
             dict of the item id keys and their estimated new price
         """
         pass
